Rename `Attribute` to `XmlAttribute` and related changes
