+++ lib/xylophone/src/core/xylophone.xmlAttribute.scala
@@ -32,6 +32,17 @@
                                                                                                   */
 package xylophone
 
-import proscenium.*
+import anticipation.*
+import contingency.*
+import gossamer.*
+import rudiments.*
+import vacuous.*
 
-case class xmlAttribute() extends StaticAnnotation
+case class XmlAttribute(node: XmlNode, attribute: Text):
+  def as[value: XmlDecoder]: value raises XmlError =
+    val attributes = Xml.normalize(node).prim match
+      case XmlAst.Element(_, _, attributes, _) => attributes
+      case _                                   => abort(XmlError(XmlError.Reason.Read))
+
+    value.read
+     (List(XmlAst.Element(XmlName(t"empty"), List(XmlAst.Textual(attributes(XmlName(attribute)))))))
